Jisang

+++ back/src/main/java/com/rdrg/back/service/implementation/PaymentServiceImplementation.java
@@ -41,6 +41,110 @@ public class PaymentServiceImplementation implements PaymentService {
     private final RentDetailRepository rentDetailRepository;
     
     @Override
+    public ResponseEntity<? super GetPaymentResponseDto> getPayment(String rentUserId) {
+
+        try {
+            boolean isExistUser = userRepository.existsByUserId(rentUserId);
+            if(!isExistUser) return ResponseDto.authenticationFailed();
+
+            DeviceRentStatusEntity reservations = paymentRepository.findTop1ByRentUserIdOrderByRentNumberDesc(rentUserId);
+            if(reservations == null) return ResponseDto.notFound();
+
+            return GetPaymentResponseDto.success(reservations);
+        } catch (Exception exception) {
+            exception.printStackTrace();
+            return ResponseDto.databaseError();
+        }
+    }
+
+    @Override
+    public ResponseEntity<? super GetPaymentListResponseDto> getAdminPaymentList() {
+        
+        try {
+            List<DeviceRentStatusEntity> deviceRentStatusEntities = paymentRepository.findByOrderByRentNumberDesc();
+            List<AdminRentItem> adminRentList = new ArrayList<>();
+
+            for (DeviceRentStatusEntity deviceRentStatusEntity: deviceRentStatusEntities) {
+                // Integer rentNumber =  deviceRentStatusEntity.getRentNumber();
+                // List<RentDetailEntity> rentDetailEntities = rentDetailRepository.findByRentNumber(rentNumber);
+                // AdminRentItem adminRentItem = new AdminRentItem(deviceRentStatusEntity, rentDetailEntities);
+                // adminRentList.add(adminRentItem);
+
+                AdminRentItem adminRentItem = createRentItem(deviceRentStatusEntity, AdminRentItem.class);
+                adminRentList.add(adminRentItem);
+            }
+            return GetAdminPaymentListResponseDto.success(adminRentList);
+        } catch (Exception exception) {
+            exception.printStackTrace();
+            return ResponseDto.databaseError();
+        }
+    }
+
+    @Override
+    public ResponseEntity<? super GetPaymentListResponseDto> getPaymentList(String rentUserId) {
+
+        try {
+            boolean isExistUser = userRepository.existsByUserId(rentUserId);
+            if(!isExistUser) return ResponseDto.authenticationFailed();
+
+            List<DeviceRentStatusEntity> deviceRentStatusEntities = paymentRepository.findByRentUserIdOrderByRentNumberDesc(rentUserId);
+            List<RentItem> rentList = new ArrayList<>();
+
+            for (DeviceRentStatusEntity deviceRentStatusEntity: deviceRentStatusEntities) {
+                // Integer rentNumber =  deviceRentStatusEntity.getRentNumber();
+                // List<RentDetailEntity> rentDetailEntities = rentDetailRepository.findByRentNumber(rentNumber);
+                // RentItem rentItem = new RentItem(deviceRentStatusEntity, rentDetailEntities);
+                // rentList.add(rentItem);
+
+                RentItem rentItem = createRentItem(deviceRentStatusEntity, RentItem.class);
+                rentList.add(rentItem);
+            }
+            return GetPaymentListResponseDto.success(rentList);
+        } catch (Exception exception) {
+            exception.printStackTrace();
+            return ResponseDto.databaseError();
+        }
+    }
+
+    @Override
+    public ResponseEntity<? super GetSearchAdminPaymentListResponseDto> getSearchAdminPaymentList(String searchWord) {
+        
+        try {
+            List<DeviceRentStatusEntity> deviceRentStatusEntities = paymentRepository.findByRentUserIdOrderByRentNumberDesc(searchWord);
+            List<AdminRentItem> adminRentList = new ArrayList<>();
+            
+            for (DeviceRentStatusEntity deviceRentStatusEntity: deviceRentStatusEntities) {
+                // Integer rentNumber =  deviceRentStatusEntity.getRentNumber();
+                // List<RentDetailEntity> rentDetailEntities = rentDetailRepository.findByRentNumber(rentNumber);
+                // AdminRentItem adminRentItem = new AdminRentItem(deviceRentStatusEntity, rentDetailEntities);
+
+                AdminRentItem adminRentItem = createRentItem(deviceRentStatusEntity, AdminRentItem.class);
+                adminRentList.add(adminRentItem);
+            }
+            return GetSearchAdminPaymentListResponseDto.success(adminRentList);
+        } catch (Exception exception) {
+            exception.printStackTrace();
+            return ResponseDto.databaseError();
+        }
+    }
+
+    @Override
+    public ResponseEntity<? super GetPaymentDetailListResponseDto> getPaymentDetailList(String rentUserId, int rentNumber) {
+        try {
+            UserEntity userEntity = userRepository.findByUserId(rentUserId);
+            if (userEntity == null) return ResponseDto.noExistUserId();
+
+            DeviceRentStatusEntity deviceRentStatusEntity = paymentRepository.findByRentNumber(rentNumber);
+            List<RentDetailEntity> rentDetailEntities = rentDetailRepository.findByRentNumber(rentNumber);
+
+            return GetPaymentDetailListResponseDto.success(deviceRentStatusEntity, rentDetailEntities);
+        } catch (Exception exception) {
+            exception.printStackTrace();
+            return ResponseDto.databaseError();
+        }
+    }
+
+    @Override
     public ResponseEntity<? super PostPaymentResponseDto> postPayment(PostPaymentRequestDto dto, String userId) {
 
         KakaoReady kakaoReady = null;
@@ -78,71 +182,47 @@ public class PaymentServiceImplementation implements PaymentService {
     }
 
     @Override
-    public ResponseEntity<? super GetPaymentResponseDto> getPayment(String rentUserId) {
+    public ResponseEntity<ResponseDto> patchRentStatus(int rentNumber, PatchRentStatusResponseDto patchRentStatusResponseDto ) {
 
         try {
-            boolean isExistUser = userRepository.existsByUserId(rentUserId);
-            if(!isExistUser) return ResponseDto.authenticationFailed();
+            DeviceRentStatusEntity deviceRentStatusEntity = paymentRepository.findByRentNumber(rentNumber);
+            if (deviceRentStatusEntity == null) return ResponseDto.noExistRentDetail();
 
-            DeviceRentStatusEntity reservations = paymentRepository.findTop1ByRentUserIdOrderByRentNumberDesc(rentUserId);
-            if(reservations == null) return ResponseDto.notFound();
+            deviceRentStatusEntity.setRentStatus(patchRentStatusResponseDto.getRentStatus());
 
-            return GetPaymentResponseDto.success(reservations);
-        } catch (Exception exception) {
-            exception.printStackTrace();
-            return ResponseDto.databaseError();
-        }
-    }
-
-    @Override
-    public ResponseEntity<? super GetPaymentListResponseDto> getPaymentList(String rentUserId) {
-
-        try {
-            boolean isExistUser = userRepository.existsByUserId(rentUserId);
-            if(!isExistUser) return ResponseDto.authenticationFailed();
-
-            List<DeviceRentStatusEntity> deviceRentStatusEntities = paymentRepository.findByRentUserIdOrderByRentNumberDesc(rentUserId);
-
-            List<RentItem> rentList = new ArrayList<>();
-
-            for (DeviceRentStatusEntity deviceRentStatusEntity: deviceRentStatusEntities) {
-                Integer rentNumber =  deviceRentStatusEntity.getRentNumber();
-                List<RentDetailEntity> rentDetailEntities = rentDetailRepository.findByRentNumber(rentNumber);
-                RentItem rentItem = new RentItem(deviceRentStatusEntity, rentDetailEntities);
-                rentList.add(rentItem);
+            if("반납 완료".equals(patchRentStatusResponseDto.getRentStatus())){
+            //     String returnPlace = deviceRentStatusEntity.getRentReturnPlace();
+            //     List<RentDetailEntity> rentDetailEntities = rentDetailRepository.findByRentNumber(rentNumber);
+            //     for (RentDetailEntity rentDetailEntity : rentDetailEntities) {
+            //         String serialNumber = rentDetailEntity.getSerialNumber();
+            //         DeviceEntity deviceEntity = deviceRepository.findBySerialNumber(serialNumber);
+            //         if (deviceEntity != null) {
+            //             deviceEntity.setPlace(returnPlace);
+            //             deviceRepository.save(deviceEntity);
+            //         }
+            //     }
+            // }
+                changePlace(deviceRentStatusEntity, rentNumber);
             }
-            
-            return GetPaymentListResponseDto.success(rentList);
-        } catch (Exception exception) {
-            exception.printStackTrace();
-            return ResponseDto.databaseError();
-        }
-    }
 
-    @Override
-    public ResponseEntity<? super GetPaymentDetailListResponseDto> getPaymentDetailList(String rentUserId, int rentNumber) {
-        try {
-            UserEntity userEntity = userRepository.findByUserId(rentUserId);
-            if (userEntity == null) return ResponseDto.noExistUserId();
-
-            DeviceRentStatusEntity deviceRentStatusEntity = paymentRepository.findByRentNumber(rentNumber);
-            
-            List<RentDetailEntity> rentDetailEntities = rentDetailRepository.findByRentNumber(rentNumber);
-
-            return GetPaymentDetailListResponseDto.success(deviceRentStatusEntity, rentDetailEntities);
+            paymentRepository.save(deviceRentStatusEntity);
 
         } catch (Exception exception) {
             exception.printStackTrace();
             return ResponseDto.databaseError();
         }
+        return ResponseDto.success();
     }
 
     @Override
-    public ResponseEntity<ResponseDto> deletePayment(int rentNumber) {
+    public ResponseEntity<ResponseDto> deletePayment(int rentNumber, String userId) {
 
         try {
             DeviceRentStatusEntity deviceRentStatusEntity = paymentRepository.findByRentNumber(rentNumber);
-            if (deviceRentStatusEntity == null) ResponseDto.noExistRentDetail();
+            if (deviceRentStatusEntity == null) return ResponseDto.noExistRentDetail();
+
+            boolean isEqual = deviceRentStatusEntity.getRentUserId().equals(userId);
+            if (!isEqual) return ResponseDto.authorizationFailed();
 
             paymentRepository.delete(deviceRentStatusEntity);
         } catch (Exception exception) {
@@ -150,81 +230,34 @@ public class PaymentServiceImplementation implements PaymentService {
             return ResponseDto.databaseError();
         }
         return ResponseDto.success();
-    }
+    }    
 
-    @Override
-    public ResponseEntity<ResponseDto> patchRentStatus(int rentNumber, PatchRentStatusResponseDto patchRentStatusResponseDto ) {
+    private <T> T createRentItem(DeviceRentStatusEntity deviceRentStatusEntity, Class<T> clazz) throws Exception {
 
-        try {
-            DeviceRentStatusEntity deviceRentStatusEntity = paymentRepository.findByRentNumber(rentNumber);
-            if (deviceRentStatusEntity == null) ResponseDto.noExistRentDetail();
+        Integer rentNumber = deviceRentStatusEntity.getRentNumber();
+        List<RentDetailEntity> rentDetailEntities = rentDetailRepository.findByRentNumber(rentNumber);
 
-            deviceRentStatusEntity.setRentStatus(patchRentStatusResponseDto.getRentStatus());
-
-            if("반납 완료".equals(patchRentStatusResponseDto.getRentStatus())){
-                String returnPlace = deviceRentStatusEntity.getRentReturnPlace();
-                List<RentDetailEntity> rentDetailEntities = rentDetailRepository.findByRentNumber(rentNumber);
-                for (RentDetailEntity rentDetailEntity : rentDetailEntities) {
-                    String serialNumber = rentDetailEntity.getSerialNumber();
-                    DeviceEntity deviceEntity = deviceRepository.findBySerialNumber(serialNumber);
-                    if (deviceEntity != null) {
-                        deviceEntity.setPlace(returnPlace);
-                        deviceRepository.save(deviceEntity);
-                }
-            }
+        if (clazz == RentItem.class) {
+            return clazz.cast(new RentItem(deviceRentStatusEntity, rentDetailEntities));
+        } else if (clazz == AdminRentItem.class) {
+            return clazz.cast(new AdminRentItem(deviceRentStatusEntity, rentDetailEntities));
         }
-
-            paymentRepository.save(deviceRentStatusEntity);
-            
-        } catch (Exception exception) {
-            exception.printStackTrace();
-            return ResponseDto.databaseError();
-        }   
-        return ResponseDto.success();
+        
+        throw new IllegalArgumentException(clazz.getName());
     }
 
-    @Override
-    public ResponseEntity<? super GetPaymentListResponseDto> getAdminPaymentList() {
-        
-        try {
-            List<DeviceRentStatusEntity> deviceRentStatusEntities = paymentRepository.findByOrderByRentNumberDesc();
+    private void changePlace(DeviceRentStatusEntity deviceRentStatusEntity, Integer rentNumber){
 
-            List<AdminRentItem> adminRentList = new ArrayList<>();
+        String returnPlace = deviceRentStatusEntity.getRentReturnPlace();
+        List<RentDetailEntity> rentDetailEntities = rentDetailRepository.findByRentNumber(rentNumber);
 
-            for (DeviceRentStatusEntity deviceRentStatusEntity: deviceRentStatusEntities) {
-                Integer rentNumber =  deviceRentStatusEntity.getRentNumber();
-                List<RentDetailEntity> rentDetailEntities = rentDetailRepository.findByRentNumber(rentNumber);
-                AdminRentItem adminRentItem = new AdminRentItem(deviceRentStatusEntity, rentDetailEntities);
-                adminRentList.add(adminRentItem);
+        for (RentDetailEntity rentDetailEntity : rentDetailEntities) {
+            String serialNumber = rentDetailEntity.getSerialNumber();
+            DeviceEntity deviceEntity = deviceRepository.findBySerialNumber(serialNumber);
+            if (deviceEntity != null) {
+                deviceEntity.setPlace(returnPlace);
+                deviceRepository.save(deviceEntity);
             }
-            
-            return GetAdminPaymentListResponseDto.success(adminRentList);
-        } catch (Exception exception) {
-            exception.printStackTrace();
-            return ResponseDto.databaseError();
-        }
-    }
-
-    @Override
-    public ResponseEntity<? super GetSearchAdminPaymentListResponseDto> getSearchAdminPaymentList(String searchWord) {
-        
-        try {
-
-            List<DeviceRentStatusEntity> deviceRentStatusEntities = paymentRepository.findByRentUserIdOrderByRentNumberDesc(searchWord);
-
-            List<AdminRentItem> adminRentList = new ArrayList<>();
-
-            for (DeviceRentStatusEntity deviceRentStatusEntity: deviceRentStatusEntities) {
-                Integer rentNumber =  deviceRentStatusEntity.getRentNumber();
-                List<RentDetailEntity> rentDetailEntities = rentDetailRepository.findByRentNumber(rentNumber);
-                AdminRentItem adminRentItem = new AdminRentItem(deviceRentStatusEntity, rentDetailEntities);
-                adminRentList.add(adminRentItem);
-            }
-            
-            return GetSearchAdminPaymentListResponseDto.success(adminRentList);
-        } catch (Exception exception) {
-            exception.printStackTrace();
-            return ResponseDto.databaseError();
         }
     }
 }
